refactor(speech): Migrate Speech component to ui directory
