+++ components/ui/speech/index.tsx
@@ -1,0 +1,5 @@
+import dynamic from "next/dynamic";
+
+export const Speech = dynamic(() => import("./speech"), {
+  ssr: false,
+});
